Moved functionality from TempHudStage to HudStage
